better libc separation (for non-wrapped case)
When not wrapping libc, be sure to load libc.so, not the
wrapped name: _ibc_so

When _not_ wrapping libc be sure to use the standard dlsym for libc
entry points (only use the table_dlsym for the library we wrap).

Also - added memcpy to the list of libc functions to look for.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>