logging / printing / debugging updates
massage logging output using more standardized macros in wrap_lib.h

make sure to chmod the logfile 0666 so that any process with any user
creds can open and write to it.

move Android logs to: /data/trace_logs
(which must be created 0777 by init)

create a handy debugging macro "lnk_dbg" which can be used in
conjunction with the LD_DEBUG environment variable. This macro accepts a
string (no formatting, just a plan const char *), and does a dlsym for
this string in the real libc DSO handle. When LD_DEBUG is set to
something greater than 1, these dlsym invocations show up in the log -
it's a handy way to debug a wrapped libc :-)

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>