bugfix! slashes in "progname" cause logfile issues
If the real libc provided "__progname" symbol contains slashes, e.g.,
the program was started from an absolute path, then the log file name
would be generated incorrectly.

This patch addresses the issue by using a local version of strrchr to
search for the last '/' character in progname. If one is found, use the
characters after it, otherwise use the whole progname.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>