Allow selective logging
In the tracer function, check for the existence of a file. If the file
exists, then allow logging to continue. If the file does not exist, then
just return and pass control back to the program.

On Android the file is: /data/enable_trace_logs
On other systems it is: /tmp/enable_trace_logs

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>