drasticaly reduct the save/restore complexity (no TLS)
Remove the TLS save/restore functionality for now. We may want this in
the future, but I was not properly initializing the TLS memory on thread
creation which was causing issues in the backtrace functionality, and
occasional SIGILLs. Because we don't need and post-function-call
tracing, we can simply push state onto the stack and then pop it back
off to make the real function call. This is much cleaner and simpler.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>