refactor should_log logic: always run wrap_special
This patch refactors the should_log logic such that the wrap_special
function is called whether or not logging is enabled. While this isn't
strictly necessary right now, it will be when dealing with signals.

Also make the inline should_log function clear the errno value to
maintain the principle of least surprise.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>