bugfix! libc wrapping always segfaulted in init_libc_iface
libc wrapping uses table_dlsym to find libc entry points, and a previous
commit added a BUG() statement on table_dlsym failure. We don't want
that failure in init_libc_iface - we expect some symbols not to be
there. This patch adds a parameter to the table_dlsym to switch between
the different functionality.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>