bugfix: give table_dlsym access to weak symbols
The objdump regex used in wraplib.sh excluded weak function symbols from
the function list used to populate real_syms.h. This caused an error
when the table_dlsym function went looking for one of the weak symbols.
While this patch resolves this particular issue, it does not mimic the
weak symbol definition, and thus may alter functionality in libraries
that depend on this... I haven't encountered any problems yet...

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>