increase backtrace speed (slightly)
This patch is more prepatory than beneficial - it removes some cruft,
but doesn't change the tracing speed by all that much.

NOTE: The log format has changed slightly. There is no longer an ending
line for the backtrace; the BT:START: line now contains the number of
subsequent lines in the trace.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>