pthread/TLS bugfixes and tweaks
This commit patches up some holes in the TLS handling during pthread
creation and fork/exec. Basically, everytime we fork or create a thread
we need to ensure that the new thread will start with its TLS data in a
zero-state. The only way to do that is to zero out our own state and
have both parties re-create the necessary data structures.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>