major refactor: support return value interposition
This is a major (an messy - sorry!) commit that adds the ability to completely
interpose on an API call while still logging. The basic concept is to keep the
function return value(s) in a small bit of TLS memory which is accessed by a
generic return function.

Currently this only supports 32-bit Android functions, but could easily be
extended to support other platforms / bit-widths.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>