bugfix! libdvm low-level interface sucks
There is at least one obscure bug in dvmGetThreadName() that causes occasional
crashes. We don't really need the java thread name, so I've just disabled it.

A more insidius offender is the dvmFillStackTraceArray() function. It takes an
array and a length as parameters and fills in the method pointers corresponding
to the Dalvik stack. However, it effectively ignores the length parameter, and
proceeds to continue writing into memory beyond the bound of your array if the
actual stack is too big for your array. Thanks guys.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>