bugfix! guard against bad/invalid unwind info
This issue shows up when there is bad/corrupt/invalid unwind information
somewhere in the middle of your stack trace. In manifests itself as an infinite
loop where the trace function gets called indefinitely with the same frame
pointer.

I put a max recursion depth counter on my tracing state structure to guard
against this, and cut off the tracing if we're not doing anymore valid work.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>