Move newline printing to bt_printf macro, dont sprintf timestamp
Use memcpy to copy the timestamp string, and move newline printing into the
bt_printf macro.

This patch also includes a subtle work-around to an annoying bug. By adding a
space after every newline (now appended in the macro), we work around a bug
where either libz or some location in my code that I haven't been able to find
overwrites the last byte with a 0. I swear it's a bug in libz...

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>