Don't call gzflush as often, use an extra gzwrite
the extra gzwrite is a hack-ish way to work around a bug where the previously
written character is overwritten by a zero. I swear this is in libz.

Occasionally you will see larger chunks of text overwritten (seemingly
haphazardly), but these are fortunately rare and can be worked around
during post-processing.

Signed-off-by: Jeremy C. Andrus <jeremya@cs.columbia.edu>